requirements.txt
numpy
#picamera
Pillow